made tiles plugin optional

+++ moodle-docker/moodle-php-apache/root/tmp/setup/plugins/download.sh
@@ -3,11 +3,13 @@
 # Load environment variables that are otherwise not visible inside this script
 source /tmp/setup/config.env 
 
-echo "Downloading Format Tiles Plugin..."
-cd /var/www/html/moodle/course/format && curl "https://moodle.org/plugins/download.php/28650/format_tiles_${PLUGIN_TILES_FORMAT_VERSION}.zip" --output tiles.zip && unzip tiles.zip && rm tiles.zip
-
 echo "Downloading Section Auto Link Filter..."
 cd /var/www/html/moodle/filter && curl "https://moodle.org/plugins/download.php/19385/filter_sectionnames_${PLUGIN_SECTION_AUTO_LINKING_VERSION}.zip" --output sectionnames.zip && unzip sectionnames.zip && rm sectionnames.zip
+
+if [ "$PLUGIN_TILES" = true ]; then
+    echo "Downloading Format Tiles Plugin..."
+    cd /var/www/html/moodle/course/format && curl "https://moodle.org/plugins/download.php/28650/format_tiles_${PLUGIN_TILES_FORMAT_VERSION}.zip" --output tiles.zip && unzip tiles.zip && rm tiles.zip
+fi 
 
 if [ "$PLUGIN_AUTOCOMPLETE" = true ]; then
     echo "Downloading plugin autocomplete..."
